// examples/todo-api/generated/todo.gen.ts
export const CreateTodoBodySchema = z.object({
  title: z.string(),
  done: z.boolean().optional(),
}).strict()
export type CreateTodoBody = z.infer<typeof CreateTodoBodySchema>

export const CreateTodoInputSchema = z.object({
  body: CreateTodoBodySchema,
}).strict()
export type CreateTodoInput = z.infer<typeof CreateTodoInputSchema>

export const TodoSchema = z.object({
  id: z.string(),
  title: z.string(),
  done: z.boolean(),
  createdAt: z.string().datetime(),
}).strict()
export type Todo = z.infer<typeof TodoSchema>

export const CreateTodoOutputSchema = z.object({
  item: TodoSchema,
}).strict()
export type CreateTodoOutput = z.infer<typeof CreateTodoOutputSchema>
// ...
export const GetTodoOutputSchema = z.object({
  item: TodoSchema,
}).strict()
export type GetTodoOutput = z.infer<typeof GetTodoOutputSchema>

export const ListTodosInputSchema = z.object({
  nextToken: z.string().optional(),
}).strict()
export type ListTodosInput = z.infer<typeof ListTodosInputSchema>

export const ListTodosOutputSchema = z.object({
  items: z.array(TodoSchema),
  nextToken: z.string().optional(),
}).strict()
export type ListTodosOutput = z.infer<typeof ListTodosOutputSchema>
// ...
export class TodoNotFound extends Error {
  readonly $statusCode = 404
  readonly $fault = 'client' as const
  constructor(message: string) {
    super(message)
    this.name = 'TodoNotFound'
    Object.setPrototypeOf(this, TodoNotFound.prototype)
  }
}

// ---- Operations interface ----

export interface TodoOperations {
  CreateTodo(input: { body: CreateTodoBody }, c?: Context<SecurityEnv>): Promise<CreateTodoOutput>
  DeleteTodo(input: { id: string }, c?: Context<SecurityEnv>): Promise<void>
  GetTodo(input: { id: string }, c?: Context<SecurityEnv>): Promise<GetTodoOutput>
  ListTodos(input: { nextToken?: string }, c?: Context<SecurityEnv>): Promise<ListTodosOutput>
}